+++ lib/widgets/components/entry/login/login_email_text_field.dart
@@ -3,10 +3,7 @@ import 'package:coffeecard/blocs/login/login_bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-// TODO Error could be an enum, e.g. LoginError.None or LoginError.InvalidEmail
 
-// FIXME Don't hide the keyboard when submitting via the keyboard action button
-// (Most likely requires a TextFormField)
 class LoginEmailTextField extends StatelessWidget {
   const LoginEmailTextField();
